+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All Notable changes to `League\Uri\Interfaces` will be documented in this file
 - `UriInterface::equals`
 - `UriInterface::toNormalizedString`
 - `UriInterface::getUser`
+- `League\Uri\IPv6\Converter::isIpv6`
 
 ### Fixed